refactor: put chat model mapper in the api.dto package
