+++ chat/src/main/kotlin/net/thechance/chat/service/model/ChatModel.kt
@@ -1,6 +1,5 @@
 package net.thechance.chat.service.model
 
-import net.thechance.chat.api.dto.ChatResponse
 import java.util.UUID
 
 data class ChatModel (
@@ -9,12 +8,3 @@ data class ChatModel (
     val requesterId : UUID,
     val id: UUID,
     )
-
-fun ChatModel.toResponse(): ChatResponse{
-    return ChatResponse(
-        id = id,
-        name = name,
-        requesterId = requesterId,
-        imageUrl = imageUrl
-    )
-}
